Refactored Entry Server into that AND Search Server
