+++ app/models/search.server.ts
@@ -1,0 +1,69 @@
+import type { Entry } from "@prisma/client"
+import { prisma } from "~/db.server"
+import { isNonPositive } from "~/utils/numberUtils"
+
+export type { Entry } from "@prisma/client"
+
+const DEFAULT_PAGE_SIZE = 100
+
+export function getEntriesByBasicTextSearchAndPage(
+  text: string,
+  page: string = "1",
+  caseSensitive: boolean = false
+) {
+  const pageNumber = parseInt(page)
+  if (isNaN(pageNumber)) {
+    throw new Error(`Page Number ("${page}") must be a number`)
+  } else if (isNonPositive(pageNumber)) {
+    throw new Error(`Page Number ("${page}") must be greater than zero`)
+  }
+
+  const skip: number = (pageNumber - 1) * DEFAULT_PAGE_SIZE
+  return getEntriesByBasicTextSearch(text, skip, undefined, caseSensitive)
+}
+
+export function getEntriesByBasicTextSearch(
+  text: string,
+  skip: number = 0,
+  take: number = 100,
+  caseSensitive: boolean = false
+) {
+  if (text.length === 0) {
+    throw new Error(`Text ("${text}") length must be greater than zero`)
+  }
+  const searchWildcard = `%${text}%`
+
+  return prisma.$queryRaw<
+    Pick<Entry, "id" | "headword">[]
+  >`SELECT id, headword FROM det_entries 
+    WHERE IF(${caseSensitive}, 
+      (headword) LIKE (${searchWildcard}), 
+      LOWER(headword) LIKE LOWER(${searchWildcard}))  
+    ORDER BY headword ASC LIMIT ${take} OFFSET ${skip}`
+}
+
+export function getSearchResultsFromMeanings(
+  text: string,
+  caseSensitive: boolean = false
+) {
+  if (text.length === 0) {
+    throw new Error(`Text ("${text}") length must be greater than zero`)
+  }
+
+  return prisma.meaning.findMany({
+    where: {
+      definition: {
+        contains: text,
+      },
+    },
+    select: {
+      entry: {
+        select: {
+          headword: true,
+        },
+      },
+      definition: true,
+      id: true,
+    },
+  })
+}
